feat(functionspace): add support for dof priority indication in tensor spaces
Introduce a `-1` element in the `shape` tuple to denote the priority of degrees
of freedom (DoFs) in arrangement. This change simplifies the specification of
tensor spaces by eliminating the separate `dof_priority` argument and instead
using the shape tuple to convey this information.

BREAKING CHANGE: The `dof_priority` constructor argument has been removed in
favor of a shape tuple with a `-1` indicator. This change impacts the way
tensor spaces are initialized and may require updates to existing code that
rely on the previous signature.